Add development requirements (requirements.dev.txt)
To be able to run the tests of datapackage developers are going to
have to install two different packages:

* nose (nosetests)
* mock (mock library for tests)

These have been added to a requirements.dev.txt file. Note that
this is not the conventional requirements.txt file, since these
are only requirements for those who hack on datapackage.