Conform to pep8 and other checkers
This adds flake8 as a checker to requirements.dev.txt. flake8 is
a wrapper around pep8, pyflakes and mccabe which can be run
against the source code to check for pep8 errors/warnings and
other errors in the source file (like unused imports).

After running flake8 against the code base a few errors have been
found and are corrected.