Convert Resource into a Specification object (dictionary)
This allows us to do validation on input if we want and the
resource can now be dropped into a json serialization since it is
already a dict. This mean resource no longer uses the descriptor
to hold its representation (it is the representation itself).

In addition, since Specification objects rely on a predetermined
dictionary of allowed keys and their types this will make it
easier to support multiple versions of the data package protocol.