Migrate datapackage.licenses to the Specification inheritance model
This required different tests and moving helper function logic to
Resource (logic is duplicated as the DataPackage class uses the
helper functions to handle licenses and the DataPackage class does
not inherit from from the Specification class).

One test is skipped because that requires DataPackage license
functionality which isn't implemented.