Add Travis-CI badget to the README
It is common to add different status badges and since we use
Travis CI for the project we can add a continuous integration badge
to the top of the README.

However the badges should not be shown on pypi which will only cause
confusion because even if there happen to be failing tests on the
master branch, the uploaded versions aren't necessarily broken (and
shouldn't be, i.e. even more strict than the master branch). The
long description reads directly from the README.rst so in order to
avoid travis badges on pypi, we pass the README.rst contents
through a filter that ignores lines that contain the travis-ci.org
http endpoint.